+++ Utils.h
@@ -21,7 +21,6 @@
 
 #if defined(STM32F10X_MD)
 #include "stm32f10x.h"
-#include <cstddef>
 #else
 #include <Arduino.h>
 #endif
@@ -32,7 +31,7 @@ uint8_t countBits32(uint32_t bits);
 
 uint8_t countBits64(uint64_t bits);
 
-uint8_t *i2str(uint8_t *dest, size_t n, int32_t x);
+uint8_t *i2str(uint8_t *dest, uint32_t n, int32_t x);
 
 #endif
